**WARN**: This is a backward incompatible change. In this commit we rewrite mos$
This is the new calling API:
```
    engine = Renderer()
    engine.render(template)
```
Where `Rendered` is the new class provided by secretary. It accepts an Enviroment
instance as its only param.

Now you pass the template when calling `render` function along with the template data.

This changes allows to reuse enviroments and/or render multiple documents using a single `Renderer` instance.