Remove bug that considers autoescaping functionality as activated if it can handle .XML files. See #41.
It really didn't matter if the autoescape function could handle XML files because secretary templates sources are not file objects. Instead, they are strings and thus lack a "filename" property necessary for environment.autoescape to work.

For now, if environment.autoescape is not True, we simply returns the template source wrapped inside an "autoescape" block.

For Jinja <2.9 environments without the autoescape extension users will have to manually escape each and every variable that may cause issues.

For now, this resolves #41.